+++ pkg/gen/input/workflows/internal/file/check_values_schema.go
@@ -3,8 +3,8 @@ package file
 import (
 	_ "embed"
 
-	"github.com/giantswarm/devctl/pkg/gen/input"
-	"github.com/giantswarm/devctl/pkg/gen/input/workflows/internal/params"
+	"github.com/giantswarm/devctl/v6/pkg/gen/input"
+	"github.com/giantswarm/devctl/v6/pkg/gen/input/workflows/internal/params"
 )
 
 //go:embed check_values_schema.yaml.template